+++ ledger/tracker_test.go
@@ -301,3 +301,16 @@ func TestTrackerDbRoundDataRace(t *testing.T) {
 	stallingTracker.cancelTasks = true
 	close(stallingTracker.produceReleaseLock)
 }
+
+func TestAccountUpdatesLedgerEvaluatorNoBlockHdr(t *testing.T) {
+	partitiontest.PartitionTest(t)
+
+	aul := &accountUpdatesLedgerEvaluator{
+		prevHeader: bookkeeping.BlockHeader{},
+		tail:       &txTail{},
+	}
+	hdr, err := aul.BlockHdr(99)
+	require.Error(t, err)
+	require.Equal(t, ledgercore.ErrNoEntry{}, err)
+	require.Equal(t, bookkeeping.BlockHeader{}, hdr)
+}
